+++ packages/core/src/lib/components/ui/calendar/Calendar.svelte
@@ -86,7 +86,7 @@
 		day?: Snippet<
 			[
 				{
-					date: Date | undefined;
+					date: Date;
 					isCurrentMonth: boolean;
 					isSelected: boolean;
 					isInRange: boolean;